bootstrap/requirements.txt
# if you just want the command-line samples experience, use this requirements file
azure-devops==5.0.0b5
python-dotenv==v0.10.2
